Move migrations to core bundle (see #2545)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2364

Commits
-------

1d6f4171 Move migrations to core bundle
20dfde35 Fix missing globals array
51fb5c72 PHPStan
5034731d Pass the filesystem as dependency and remove the Contao 3 migrations
