+++ installation-bundle/src/DependencyInjection/ContaoInstallationExtension.php
@@ -29,6 +29,5 @@ class ContaoInstallationExtension extends Extension
         $loader->load('commands.yml');
         $loader->load('listener.yml');
         $loader->load('services.yml');
-        $loader->load('migrations.yml');
     }
 }
